Revert "feat: backport user task PIM to stable/8.8"
