+++ qa/util/src/main/java/io/camunda/qa/util/cluster/TestRestTasklistClient.java
@@ -8,11 +8,8 @@
 package io.camunda.qa.util.cluster;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import io.camunda.client.CredentialsProvider;
-import io.camunda.tasklist.webapp.api.rest.v1.entities.TaskSearchResponse;
 import io.camunda.zeebe.util.CloseableSilently;
 import java.net.URI;
 import java.net.http.HttpClient;
@@ -29,8 +26,7 @@ import java.util.function.Consumer;
 
 public class TestRestTasklistClient implements CloseableSilently {
 
-  public static final ObjectMapper OBJECT_MAPPER =
-      new ObjectMapper().registerModule(new JavaTimeModule());
+  public static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
 
   private final URI endpoint;
   private final HttpClient httpClient;
@@ -38,12 +34,13 @@ public class TestRestTasklistClient implements CloseableSilently {
 
   public TestRestTasklistClient(final URI endpoint) {
     this.endpoint = endpoint;
-    httpClient = HttpClient.newHttpClient();
+    this.httpClient = HttpClient.newHttpClient();
   }
 
   public TestRestTasklistClient(final URI endpoint, final CredentialsProvider credentialsProvider) {
     this(endpoint);
-    authenticationApplier = TestRestOperateClient.applyCredentialsProvider(credentialsProvider);
+    this.authenticationApplier =
+        TestRestOperateClient.applyCredentialsProvider(credentialsProvider);
   }
 
   TestRestTasklistClient(
@@ -52,7 +49,7 @@ public class TestRestTasklistClient implements CloseableSilently {
       final String username,
       final String password) {
     this(endpoint);
-    authenticationApplier = applyBasicAuthenticationHeader(username, password);
+    this.authenticationApplier = applyBasicAuthenticationHeader(username, password);
   }
 
   public HttpResponse<String> searchRequest(final String path, final String body) {
@@ -73,12 +70,6 @@ public class TestRestTasklistClient implements CloseableSilently {
         Optional.ofNullable(processInstanceKey)
             .map(a -> mapToRequestBody("processInstanceKey", a))
             .orElse(null));
-  }
-
-  public TaskSearchResponse[] searchAndParseTasks(final Long processInstanceKey)
-      throws JsonProcessingException {
-    return OBJECT_MAPPER.readValue(
-        searchTasks(processInstanceKey).body(), TaskSearchResponse[].class);
   }
 
   public HttpResponse<String> createProcessInstanceViaPublicForm(final String processDefinitionId) {
