+++ tasklist/importer-870/src/main/java/io/camunda/tasklist/zeebeimport/v870/record/value/JobRecordValueImpl.java
@@ -35,7 +35,6 @@ public class JobRecordValueImpl extends RecordValueWithPayloadImpl implements Jo
   private JobListenerEventType jobListenerEventType;
   private Set<String> changedAttributes;
   private Set<String> tags;
-  private boolean jobToUserTaskMigration;
 
   public JobRecordValueImpl() {}
 
@@ -210,11 +209,6 @@ public class JobRecordValueImpl extends RecordValueWithPayloadImpl implements Jo
   @Override
   public Set<String> getTags() {
     return tags;
-  }
-
-  @Override
-  public boolean isJobToUserTaskMigration() {
-    return false;
   }
 
   @Override
