+++ zeebe/exporters/camunda-exporter/src/main/java/io/camunda/exporter/handlers/UserTaskJobBasedHandler.java
@@ -86,8 +86,7 @@ public class UserTaskJobBasedHandler implements ExportHandler<TaskEntity, JobRec
   @Override
   public boolean handlesRecord(final Record<JobRecordValue> record) {
     return SUPPORTED_INTENTS.contains(record.getIntent())
-        && record.getValue().getType().equals(Protocol.USER_TASK_JOB_TYPE)
-        && !record.getValue().isJobToUserTaskMigration();
+        && record.getValue().getType().equals(Protocol.USER_TASK_JOB_TYPE);
   }
 
   @Override
