+++ zeebe/exporters/elasticsearch-exporter/src/main/resources/zeebe-record-job-batch-template.json
@@ -104,9 +104,6 @@
                 "tags": {
                   "type": "keyword"
                 },
-                "jobToUserTaskMigration": {
-                  "type": "boolean"
-                },
                 "result": {
                   "properties": {
                     "type": {
