+++ zeebe/exporters/elasticsearch-exporter/src/main/resources/zeebe-record-job-template.json
@@ -85,9 +85,6 @@
             "tags": {
               "type": "keyword"
             },
-            "jobToUserTaskMigration": {
-              "type": "boolean"
-            },
             "result": {
               "properties": {
                 "type": {
