+++ zeebe/exporters/opensearch-exporter/src/test/java/io/camunda/zeebe/exporter/opensearch/OpensearchExporterIT.java
@@ -417,7 +417,6 @@ final class OpensearchExporterIT {
                   ImmutableJobRecordValue.builder()
                       .from(((ImmutableJobRecordValue) record.getValue()))
                       .withTags(List.of())
-                      .withJobToUserTaskMigration(false)
                       .resultBuilder(null)
                       .build();
               case PROCESS_INSTANCE ->
@@ -439,7 +438,6 @@ final class OpensearchExporterIT {
                                           ImmutableJobRecordValue.builder()
                                               .from(job)
                                               .withTags(List.of())
-                                              .withJobToUserTaskMigration(false)
                                               .resultBuilder(null)
                                               .build())
                                   .toList())
